Start work on refactoring Added flow config
