+++ bot/.babelrc
@@ -1,4 +1,4 @@
 {
-    "presets": ["react", "es2015", "es2017"],
+    "presets": ["flow", "react", "es2015", "es2017"],
     "plugins": ["transform-object-rest-spread", "transform-runtime"],
 }
