+++ bot/package.json
@@ -10,7 +10,7 @@
   "private": true,
   "devDependencies": {
     "autobind-decorator": "^1.3.3",
-    "babel-cli": "^6.24.1",
+    "babel-cli": "^6.26.0",
     "babel-core": "^6.24.1",
     "babel-eslint": "^7.1.1",
     "babel-loader": "^7.1.1",
@@ -23,6 +23,7 @@
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-es2017": "^6.24.1",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.3.13",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.5.0",
@@ -39,6 +40,7 @@
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.8.0",
+    "flow-bin": "^0.55.0",
     "html-webpack-plugin": "^2.29.0",
     "pretty-hrtime": "^1.0.1",
     "proxy-middleware": "^0.15.0",
@@ -54,6 +56,7 @@
   },
   "dependencies": {
     "browser-request": "^0.3.3",
+    "fuse.js": "^3.1.0",
     "jquery": "^3.2.1",
     "lodash": "^4.17.4",
     "material-ui": "^0.16.7",
