+++ bot/src/app/components/shops/supreme/pages/Billing.jsx
@@ -9,7 +9,7 @@ import {
 } from 'redux-form-material-ui';
 import Styles from '../../../../constants/Styles';
 import * as Utils from '../../../../constants/Utils';
-import * as Validators from '../../../../constants/FormValidators';
+import * as Validators from '../../../../utils/FormValidators';
 import * as menus from '../../../../constants/Menus';
 
 function getStatesForCountry(country) {
