+++ bot/src/app/utils/FormValidators.js
@@ -1,4 +1,4 @@
-import * as Helpers from '../utils/Helpers';
+import * as Helpers from './Helpers';
 
 export function email(input) {
   const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
