+++ bot/src/app/utils/Helpers.js
@@ -1,5 +1,4 @@
 import moment from 'moment';
-import { findBestMatch } from '../../extension/helpers';
 
 export function isObjectEmpty(obj) {
   return Object.keys(obj).length === 0 && obj.constructor === Object;
