+++ bot/src/extension/content/index.js
@@ -1,0 +1,11 @@
+import SupremeExtension from './supreme/index';
+import StorageService from '../../services/StorageService';
+
+
+async function start() {
+  const profile = await StorageService.getCurrentProfileSettings();
+  const supremeExtension = new SupremeExtension();
+  supremeExtension.start(profile);
+}
+
+start();
