+++ bot/src/extension/content/supreme/SupremeManager.js
