+++ bot/src/extension/content/supreme/helpers.js
