+++ bot/src/extension/content/supreme/notification.js
