+++ bot/src/extension/content/supreme/processors/baseProcessor.js
