+++ bot/src/extension/content/supreme/processors/productProcessor.js
