+++ bot/src/services/ChromeService.js
@@ -1,0 +1,5 @@
+export default class ChromeService {
+  static openTab(url) {
+
+  }
+}
