+++ bot/src/services/supreme/ProductWatcherService.js
@@ -1,0 +1,11 @@
+import StorageService from '../StorageService';
+
+export default class ProductWatcherService {
+  static async getProducts() {
+    return await StorageService.getItem('products');
+  }
+
+  static setProducts(products) {
+    return StorageService.setItem('products', products);
+  }
+}
